feat: implementation hook useEffect for function component useEscClose.
