+++ src/components/hooks/useEscClose.ts
@@ -1,0 +1,22 @@
+import { useEffect, useContext } from "react";
+import ModalContext from "./ModalContext";
+
+
+const useEscClose = () => {
+    const { closeModal } = useContext(ModalContext);
+
+    useEffect(() => {
+        // function handleEscKey now always will be closed form
+        const handleEscKey = (event: { keyCode: number }) => {
+          if (event.keyCode === 27) {
+            closeModal();
+          }
+        };
+        document.addEventListener("keydown", handleEscKey);
+        return () => {
+          document.removeEventListener("keydown", handleEscKey);
+        };
+    }, [closeModal]); // Перерегистрация обработчика при изменении closeModal
+}
+
+export default useEscClose;
