+++ src/layout/header/mobileMenu/MobileMenu.tsx
@@ -1,9 +1,10 @@
 import { BurgerButton } from "../../../components/menu/burger/BurgerButton";
 import useModalToggle from "../../../components/hooks/useModalToggle";
 import { S } from "../../../components/menu/headerMenu/HeaderMenu.styles";
-import ModalContext from "../../../components/functions/ModalContext";
+import ModalContext from "../../../components/hooks/ModalContext";
 import { Menu } from "../../../components/menu/Menu";
-import React, { useContext, useEffect } from "react";
+import React, { useContext } from "react";
+import useEscClose from "../../../components/hooks/useEscClose";
 
 type MobileMenuPopupProps = {
   isOpen: boolean;
@@ -13,18 +14,7 @@ export const MobileMenu: React.FC<MobileMenuPopupProps> = () => {
   const {isModalOpen: isMenuOpen } = useModalToggle();
   const { modalOpen, openModal, closeModal } = useContext(ModalContext);
 
-  useEffect(() => {
-    // function handleEscKey now always will be closed form
-    const handleEscKey = (event: { keyCode: number }) => {
-      if (event.keyCode === 27) {
-        closeModal();
-      }
-    };
-    document.addEventListener("keydown", handleEscKey);
-    return () => {
-      document.removeEventListener("keydown", handleEscKey);
-    };
-  }); //   }, []); в данном контексте передавать зависимость нет необходимости дабы избежать бесконечного перерендера самого себя.
+  useEscClose();
 
   return (
     <S.MobileMenu>
